게시글 기능 :: 추가

+++ src/main/java/com/example/demu/domain/post/service/DeletePostService.java
@@ -1,7 +1,11 @@
 package com.example.demu.domain.post.service;
 
+import com.example.demu.domain.auth.facade.UserFacade;
 import com.example.demu.domain.post.domain.Post;
 import com.example.demu.domain.post.domain.repository.PostRepository;
+import com.example.demu.domain.post.exception.CannotModifyFeedException;
+import com.example.demu.domain.post.facade.PostFacade;
+import com.example.demu.domain.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,11 +15,18 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional
 
 public class DeletePostService {
+
     private final PostRepository postRepository;
+    private final UserFacade userFacade;
+    private final PostFacade postFacade;
 
     public void deletePost(Long id){
-        Post post = postRepository.findById(id)
-                .orElseThrow(RuntimeException::new);
+        User user = userFacade.CurrentUser();
+        Post post = postFacade.getPost(id);
+
+        if(!user.equals(post.getUser())) {
+            throw CannotModifyFeedException.EXCEPTION;
+        }
 
         postRepository.delete(post);
     }
